test-requirements.txt
ansible-lint==5.4.0
molecule>=3.4.0
molecule-docker>=0.2.4
docker>=5.0.0
yamllint==1.26.3
flake8==3.9.2
pytest==6.2.5
pytest-testinfra==6.4.0
tox==3.24.3
tox-gh-actions==2.6.0
